+++ code/Samples/NativeTexture/NativeTexture.cc
@@ -177,11 +177,11 @@ NativeTextureApp::computeMVP(const glm::vec3& pos) {
 AppState::Code
 NativeTextureApp::notSupported() {
     const char* msg = "This demo needs GL\n";
-    uint8_t x = (Gfx::DisplayAttrs().FramebufferWidth/16 - strlen(msg))/2;
-    uint8_t y = Gfx::DisplayAttrs().FramebufferHeight/16/2;
+    int x = (Gfx::DisplayAttrs().FramebufferWidth/16 - int(strlen(msg)))/2;
+    int y = Gfx::DisplayAttrs().FramebufferHeight/16/2;
     Gfx::ApplyDefaultRenderTarget(ClearState::ClearAll(glm::vec4(0.5f, 0.0f, 0.0f, 1.0f)));
     Dbg::SetTextScale(glm::vec2(2.0f, 2.0f));
-    Dbg::CursorPos(x, y);
+    Dbg::CursorPos(uint8_t(x), uint8_t(y));
     Dbg::Print(msg);
     Dbg::DrawTextBuffer();
     Gfx::CommitFrame();
